refactor: authConfig 스플리팅
